refactor: chatroom 내부로 로직 이동
